requirements.txt
## The following requirements were added by pip freeze:
argparse==1.3.0
backports.ssl-match-hostname==3.4.0.2
bower.py==1.0
certifi==2015.4.28
coverage==3.7.1
coveralls==0.5
decorator==3.4.0
docopt==0.6.2
docutils==0.12
futures==3.0.3
github3.py==0.9.4
importlib==1.0.3
invoke==0.11.1
itsdangerous==0.24
Jinja2==2.7.3
lockfile==0.12.2
logging==0.4.9.6
lxml==3.4.1
MarkupSafe==0.23
networkx==1.9.1
nltk==3.0.1
py==1.4.26
pymongo==3.0.3
pytest==2.6.4
python-daemon==2.1.0
PyYAML==3.11
requests==2.7.0
semantic-version==2.4.2
tornado==4.2.1
uritemplate.py==0.3.0
Werkzeug==0.10.1
wheel==0.24.0
